Adding a FLY_ENV variable so I can politely step around the Fly environment startup

+++ start.sh
@@ -5,21 +5,27 @@
 # run, which is why this file exists in the first place.
 # Learn more: https://community.fly.io/t/sqlite-not-getting-setup-properly/4386
 
-set -ex
+# Fly environment is enabled by default (built into the Docker image)
+# set it to anything other than enabled to disable...
+if [ "$FLY_ENV" = "enabled" ]; then
+  echo "Fly environment enabled"
+  set -ex
 
-# Setup 512MB of space for swap and set permissions and turn on swapmode
-fallocate -l 512M /swapfile
-chmod 0600 /swapfile
-mkswap /swapfile
-echo 10 > /proc/sys/vm/swappiness
-swapon /swapfile
+  # Setup 512MB of space for swap and set permissions and turn on swapmode
+  fallocate -l 512M /swapfile
+  chmod 0600 /swapfile
+  mkswap /swapfile
+  echo 10 > /proc/sys/vm/swappiness
+  swapon /swapfile
 
-# Run migrations
-npx prisma migrate deploy
+  # Run migrations
+  npx prisma migrate deploy
 
-# Turn off swap mode and remove swap directory
-swapoff /swapfile
-rm /swapfile
+  # Turn off swap mode and remove swap directory
+  swapoff /swapfile
+  rm /swapfile
+
+fi
 
 # Finally start the app
 npm run start
